extract common api calls
