+++ app/src/main/java/org/eyeseetea/malariacare/data/remote/api/PullDhisApiDataSource.java
@@ -5,13 +5,13 @@ import android.util.Log;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import okhttp3.Response;
 
 import org.eyeseetea.malariacare.data.database.iomodules.dhis.importer.models.EventExtended;
 import org.eyeseetea.malariacare.data.database.model.UserDB;
 import org.eyeseetea.malariacare.data.database.utils.PreferencesState;
 import org.eyeseetea.malariacare.data.database.utils.Session;
 import org.eyeseetea.malariacare.domain.exception.ClosedUserDateNotFoundException;
-import org.eyeseetea.malariacare.domain.exception.PullApiParsingException;
 import org.eyeseetea.malariacare.utils.DateParser;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -20,15 +20,10 @@ import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
+import static org.eyeseetea.malariacare.data.remote.api.OkHttpClientDataSource.executeCall;
+import static org.eyeseetea.malariacare.data.remote.api.OkHttpClientDataSource.parseResponse;
 
 public class PullDhisApiDataSource {
-
-    public static final MediaType JSON = MediaType.parse("application/json; charset=utf-8");
 
     private static final String DHIS_PULL_API="/api/";
 
@@ -173,74 +168,4 @@ public class PullDhisApiDataSource {
         return PullDhisApiDataSource.pullQuarantineEvents(url);
     }
 
-
-    /**
-     * Call to DHIS Server
-     * @param data
-     * @param method
-     * @param url
-     */
-    public static Response executeCall(JSONObject data, String url, String method) throws IOException {
-        final String DHIS_URL=PreferencesState.getInstance().getServer().getUrl() + url.replace(" ", "%20");
-
-        Log.d(TAG, "executeCall Url" + DHIS_URL + "");
-
-        BasicAuthenticator basicAuthenticator = new BasicAuthenticator();
-        OkHttpClient client = UnsafeOkHttpsClientFactory.getUnsafeOkHttpClient(basicAuthenticator);
-
-        Request.Builder builder = new Request.Builder()
-                .header(basicAuthenticator.AUTHORIZATION_HEADER, basicAuthenticator.getCredentials())
-                .url(DHIS_URL);
-
-        switch (method){
-            case "POST":
-                RequestBody postBody = RequestBody.create(JSON, data.toString());
-                builder.post(postBody);
-                break;
-            case "PUT":
-                RequestBody putBody = RequestBody.create(JSON, data.toString());
-                builder.put(putBody);
-                break;
-            case "PATCH":
-                RequestBody patchBody = RequestBody.create(JSON, data.toString());
-                builder.patch(patchBody);
-                break;
-            case "GET":
-                builder.get();
-                break;
-        }
-
-        Request request = builder.build();
-        Response response = client.newCall(request).execute();;
-        if (!response.isSuccessful()) {
-            Log.e(TAG, "pushData (" + response.code() + "): " + response.body().string());
-            throw new IOException(response.message());
-        }
-        return response;
-    }
-
-    /**
-     * Call to DHIS Server
-     * @param url
-     * @param method
-     */
-    public static Response executeCall(String url, String method) throws IOException {
-        return executeCall(null, url, method);
-    }
-
-    private static JsonNode parseResponse(String responseData)throws Exception{
-        try{
-            JSONObject jsonResponse=new JSONObject(responseData);
-            Log.i("JsonCommonParser", "parseResponse: " + jsonResponse);
-            ObjectMapper objectMapper = new ObjectMapper();
-            String jsonString = jsonResponse.toString();
-            try {
-                return objectMapper.readValue(jsonString, JsonNode.class);
-            }catch(Exception ex){
-                throw new PullApiParsingException();
-            }
-        }catch(Exception ex){
-            throw new PullApiParsingException();
-        }
-    }
 }
